rename test suite file
